fix(sdk): address verified Codex review findings — Events() buffering, filter kinds, make fix
External Codex review of PR #434 produced 8 findings; each was verified
against source and git history before acting. Fixed here (4):

- stream.go: Events() channel buffers from construction again, restoring
  TS parity (controller.ts buffers unconditionally). The chanRequested
  opt-in gate (round 1) silently discarded events emitted before the
  first Events() call; the dropLogOnce once-only drop log it was added
  alongside is kept. New test pins the buffering behavior.
- stream.go: toFloat64 handles all int/uint kinds; a codegen-typed
  narrow/unsigned operand in an ordered .Where() filter silently dropped
  matching live events. Test table extended.
- stream.go: scanner-cap comment cited the wrong bound — the binding
  per-event ceiling is the embedded NATS 1 MiB MaxPayload, not the
  16 MiB HTTP ingest cap.
- types.go: MaxRowsToRead *int -> *int64, matching the server's wire
  type (internal/policy: int64, can exceed 2^31).
- Makefile: fix-go ran gofumpt/goimports via `go tool` inside the nested
  module, which has no tool directives — `make fix` failed with
  "go: no such tool" (regression from round 1). Formatters now run from
  the root module against clients/go; tidy + golangci-lint keep their
  nested cd.

Remaining findings triaged to issues (drafted separately): failed-backfill
buffer flush (Go+TS), LIKE case semantics (Go+TS+server), Nullable+DEFAULT
explicit-null collapse (codegen).

+++ clients/go/types.go
@@ -175,7 +175,7 @@ type RolePermissions struct {
 	DeniedAggregations  []string                `json:"denied_aggregations,omitempty"`
 	MaxRows             *int                    `json:"max_rows,omitempty"`
 	MaxExecutionTime    any                     `json:"max_execution_time,omitempty"`
-	MaxRowsToRead       *int                    `json:"max_rows_to_read,omitempty"`
+	MaxRowsToRead       *int64                  `json:"max_rows_to_read,omitempty"`
 	MaxMemoryUsage      any                     `json:"max_memory_usage,omitempty"`
 }
